DockerDateien/requirements.txt
paho-mqtt
psycopg2-binary
requests
streamlit
plotly
pandas
python-dotenv
sqlalchemy